selftests/mm: on-fault-limit: run test without root privileges otherwise skip
The mmap() respects rlimit only for normal users.  This test should be run
as normal user, without root privileges.  Also add back the sudo -u nobody
as run_vmtests.sh is run as root most of the times.  Skip the test instead
if sudo isn't present to lower the privileges.

Link: https://lkml.kernel.org/r/20240201130538.1404897-1-usama.anjum@collabora.com
Fixes: b6221771d468 ("selftests/mm: run_vmtests: remove sudo and conform to tap")
Signed-off-by: Muhammad Usama Anjum <usama.anjum@collabora.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>