mm: memcg: fix struct memcg_vmstats_percpu size and alignment
Commit da10d7e140196 ("mm: memcg: optimize parent iteration in
memcg_rstat_updated()") added two additional pointers to the end of struct
memcg_vmstats_percpu with CACHELINE_PADDING to put them in a separate
cacheline.  This caused the struct size to increase from 1200 to 1280 on
my config (80 extra bytes instead of 16).

Upon revisiting, the relevant struct members do not need to be on a
separate cacheline, they just need to fit in a single one.  This is a
percpu struct, so there shouldn't be any contention on that cacheline
anyway.  Move the members to the beginning of the struct and make sure the
struct itself is cacheline aligned.  Add a comment about the members that
need to fit together in a cacheline.

The struct size is now 1216 on my config with this change.

Link: https://lkml.kernel.org/r/20240203044612.1234216-1-yosryahmed@google.com
Fixes: da10d7e14019 ("mm: memcg: optimize parent iteration in memcg_rstat_updated()")
Signed-off-by: Yosry Ahmed <yosryahmed@google.com>
Reported-by: Greg Thelen <gthelen@google.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Michal Hocko <mhocko@kernel.org>
Cc: Muchun Song <muchun.song@linux.dev>
Cc: Roman Gushchin <roman.gushchin@linux.dev>
Cc: Shakeel Butt <shakeelb@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>