mm: document memalloc_noreclaim_save() and memalloc_pin_save()
memalloc_noreclaim_save() currently has no documentation comment, so the
implications of its usage are not obvious.  Namely that it not only
prevents entering reclaim (as the name suggests), but also allows using
all memory reserves and thus should be only used in contexts that are
allocating memory to free memory.  This may lead to new improper usages
being added.

Thus add a documenting comment, based on the description of
__GFP_MEMALLOC. While at it, also document memalloc_pin_save() so that
all the memalloc_ scopes are documented. In the comments describing the
relevant PF_MEMALLOC flags, refer to their scope setting functions.

Link: https://lkml.kernel.org/r/20240212182950.32730-2-vbabka@suse.cz
Signed-off-by: Vlastimil Babka <vbabka@suse.cz>
Cc: Kent Overstreet <kent.overstreet@gmail.com>
Cc: Matthew Wilcox (Oracle) <willy@infradead.org>
Cc: Mel Gorman <mgorman@techsingularity.net>
Cc: Michal Hocko <mhocko@kernel.org>
Cc: Mike Rapoport <rppt@linux.ibm.com>
Cc: Pasha Tatashin <pasha.tatashin@soleen.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>