kfifo: add kfifo_out_linear{,_ptr}()
These are helpers which are going to be used in the serial layer.  We need
a wrapper around kfifo which provides us with a tail (sometimes "tail"
offset, sometimes a pointer) to the kfifo data.  And which returns count
of available data -- but not larger than to the end of the buffer (hence
_linear in the names).  I.e.  something like CIRC_CNT_TO_END() in the
legacy circ_buf.

This patch adds such two helpers.

Link: https://lkml.kernel.org/r/20240223115155.8806-4-jirislaby@kernel.org
Signed-off-by: Jiri Slaby (SUSE) <jirislaby@kernel.org>
Cc: Stefani Seibold <stefani@seibold.net>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>