Docs/mm/damon/design: remove the details for pageout as paddr doesn't use MADV_PAGEOUT
The doc needs fix.  As only in the case of virtual address, we are calling
madvise() with MADV_PAGEOUT.  But in the case of physical address, we are
calling reclaim_pages() directly.  MADV_PAGEOUT on virtual address is much
more aggresive to reclaim memory compared to reclaim_pages() on paddr
region.  This patch removes the details so that the description can apply
to both cases.  And we don't need to couple with the implementation
details.

Link: https://lkml.kernel.org/r/20240224224751.4673-1-21cnbao@gmail.com
Signed-off-by: Barry Song <v-songbaohua@oppo.com>
Reviewed-by: SeongJae Park <sj@kernel.org>
Cc: Minchan Kim <minchan@kernel.org>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>