arch/um: fix forward declaration for vmalloc
Patch [1] replaced vmalloc() function with a new definition but it did not
adjust the forward declaration used in UML architecture.  Change it to act
as before.

Note that this prevents the vmalloc() allocations in __wrap_malloc() from
being accounted.  If accounting here is critical, we will have to remove
this forward declaration and include vmalloc.h, however that would pull in
more dependencies and would require introducing more architecture-specific
headers, like asm/bug.h, asm/rwonce.h, etc.  This is likely the reason why
this forward declaration was introduced in the first place.

[1] https://lore.kernel.org/all/20240321163705.3067592-31-surenb@google.com/

Link: https://lkml.kernel.org/r/20240326073750.726636-1-surenb@google.com
Fixes: 576477564ede ("mm: vmalloc: enable memory allocation profiling")
Signed-off-by: Suren Baghdasaryan <surenb@google.com>
Reported-by: SeongJae Park <sj@kernel.org>
Closes: https://lore.kernel.org/all/20240323180506.195396-1-sj@kernel.org/
Tested-by: SeongJae Park <sj@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>